requirements.txt
Flask == 2.0.1
Flask_SQLAlchemy == 2.5.1
PyMySQL == 1.0.2
alembic == 1.7.4
